requirements.txt
liac-arff
matplotlib
networkx
numpy==1.26.4
scikit-learn
scikit-learn-extra
seaborn